fix unsub

+++ bin/admin.pl
@@ -68,6 +68,8 @@ if ($opt{products}) {
 	my $pid = $dbh->selectrow_array(q{SELECT id FROM product WHERE code = ?}, undef, $opt{unsub});
 	my $uid = $dbh->selectrow_array(q{SELECT id FROM user WHERE email = ?},   undef, $opt{email});
 	print "PID: $pid  UID: $uid\n";
+	die "Could not find product" if not $pid;
+	die "Could not find User" if not $uid;
 	$dbh->do(q{DELETE FROM subscription WHERE uid=? AND pid=?}, undef, $uid, $pid);
 	show_people($opt{email});
 } elsif ($opt{list}) {
